fix multiple getters panic
